Unwanted automatic orById call
Problem: If we automatically call function orById (for example in adt/ajax-select -> DynamicSelect if we pass whole QO) on QO that has no other where filters, so we want to include all, orById cannot be called otherwise it will filter only passed ID and not all records.

Solution: orById filter is called in doCreateBasicQuery after the apply of all other filters with check if there is something in WHERE part of SQL. Then there is a certainty that orById will be called at the end of all filters and at the same time it will not be called if other filters are not applied.