Superiority of byId filter
byId filter must be superior to orById filter. In result we should have ((standard filters) OR {orById filter}) AND {byIdFilter}.

The problem with original implementation was that if we were editing entity through ajaxMultiSelect in EntityForm, if we removed an item from input and saved the form. Removed item didn't remove from entity, because it was included additionally in InvalidSetValueTrait where orById filter was called and was superior to all other filters.